:construction: chore: resolver placeholder

+++ server/resolver/Task.resolver.ts
@@ -1,5 +1,4 @@
 import { Resolver, Query, Arg, Mutation } from "type-graphql";
-import { Repository, Transaction, TransactionRepository } from "typeorm";
 
 import Task from "../entity/Task";
 
@@ -9,10 +8,12 @@ import {
   TaskStatus,
 } from "../graphql/Common";
 import {
+  TaskQueryInput,
   TaskCreateInput,
   TaskUpdateInput,
   TaskRelationsInput,
   getTaskRelations,
+  TaskRelation,
 } from "../graphql/Task";
 import { DifficultyLevel } from "../graphql/Public";
 
@@ -21,6 +22,8 @@ import ExecutorService from "../service/Executor.service";
 import SubstanceService from "../service/Substance.service";
 
 import { RESPONSE_INDICATOR } from "../utils/constants";
+
+import { generatePagination } from "../utils/helper";
 
 @Resolver((of) => Task)
 export default class TaskResolver {
@@ -34,16 +37,14 @@ export default class TaskResolver {
   async QueryAllTasks(
     @Arg("pagination", { nullable: true })
     pagination: PaginationOptions,
+
     @Arg("relations", (type) => TaskRelationsInput, { nullable: true })
     relationOptions: Partial<TaskRelationsInput> = {}
   ): Promise<TaskStatus> {
     try {
-      const queryPagination = (pagination ?? {
-        cursor: 0,
-        offset: 20,
-      }) as Required<PaginationOptions>;
+      const queryPagination = generatePagination(pagination);
+      const relations: TaskRelation[] = getTaskRelations(relationOptions);
 
-      const relations = getTaskRelations(relationOptions);
       const res = this.taskService.getAllTasks(queryPagination, relations);
 
       return new StatusHandler(true, RESPONSE_INDICATOR.SUCCESS, res);
@@ -54,12 +55,13 @@ export default class TaskResolver {
 
   @Query(() => TaskStatus, { nullable: false, description: "基于ID获取任务" })
   async QueryTaskByID(
-    @Arg("taskId") taskId: string,
+    @Arg("taskId") taskId: number,
+
     @Arg("relations", (type) => TaskRelationsInput, { nullable: true })
     relationOptions: Partial<TaskRelationsInput> = {}
   ): Promise<TaskStatus> {
     try {
-      const relations = getTaskRelations(relationOptions);
+      const relations: TaskRelation[] = getTaskRelations(relationOptions);
       const res = await this.taskService.getOneTaskById(taskId, relations);
 
       if (!res) {
@@ -74,21 +76,81 @@ export default class TaskResolver {
 
   @Query(() => TaskStatus, {
     nullable: false,
+    description: "基于条件获取单个任务",
+  })
+  async QueryTaskByConditions(
+    @Arg("taskQueryParams") param: TaskQueryInput,
+
+    @Arg("relations", (type) => TaskRelationsInput, { nullable: true })
+    relationOptions: Partial<TaskRelationsInput> = {}
+  ): Promise<TaskStatus> {
+    try {
+      const relations: TaskRelation[] = getTaskRelations(relationOptions);
+
+      const res = await this.taskService.getOneTaskByConditions(
+        param,
+        relations
+      );
+
+      return new StatusHandler(true, RESPONSE_INDICATOR.SUCCESS, [res]);
+    } catch (error) {
+      return new StatusHandler(false, JSON.stringify(error), []);
+    }
+  }
+
+  @Query(() => TaskStatus, {
+    nullable: false,
+    description: "基于条件获取多个任务",
+  })
+  async QueryTasksByConditions(
+    @Arg("taskQueryParams") param: TaskQueryInput,
+
+    @Arg("pagination", { nullable: true })
+    pagination: PaginationOptions,
+
+    @Arg("relations", (type) => TaskRelationsInput, { nullable: true })
+    relationOptions: Partial<TaskRelationsInput> = {}
+  ): Promise<TaskStatus> {
+    try {
+      const queryPagination = generatePagination(pagination);
+
+      const relations: TaskRelation[] = getTaskRelations(relationOptions);
+
+      const res = await this.taskService.getTasksByConditions(
+        param,
+        queryPagination,
+        relations
+      );
+
+      return new StatusHandler(true, RESPONSE_INDICATOR.SUCCESS, res);
+    } catch (error) {
+      return new StatusHandler(false, JSON.stringify(error), []);
+    }
+  }
+
+  @Query(() => TaskStatus, {
+    nullable: false,
     description: "查询执行者当前被分配的任务",
   })
   async QueryExecutorTasks(
-    @Arg("uid") uid: string,
+    @Arg("uid") uid: number,
+
+    @Arg("pagination", { nullable: true })
+    pagination: PaginationOptions,
+
     @Arg("relations", (type) => TaskRelationsInput, { nullable: true })
     relationOptions: Partial<TaskRelationsInput> = {}
   ) {
     try {
+      const queryPagination = generatePagination(pagination);
+
       const executor = await this.executorService.getOneExecutorById(uid);
 
       if (!executor) {
         return new StatusHandler(false, RESPONSE_INDICATOR.NOT_FOUND, []);
       }
 
-      const relations = getTaskRelations(relationOptions);
+      const relations: TaskRelation[] = getTaskRelations(relationOptions);
 
       const res = await this.taskService.getTasksByConditions(
         {
@@ -96,7 +158,8 @@ export default class TaskResolver {
             uid,
           },
         },
-        [...relations, "assignee"]
+        queryPagination,
+        relations
       );
 
       return new StatusHandler(true, RESPONSE_INDICATOR.SUCCESS, res);
@@ -106,7 +169,7 @@ export default class TaskResolver {
   }
 
   @Mutation(() => TaskStatus, { nullable: false, description: "变更任务状态" })
-  async ToggleTaskStatus(@Arg("taskId") taskId: string): Promise<TaskStatus> {
+  async ToggleTaskStatus(@Arg("taskId") taskId: number): Promise<TaskStatus> {
     try {
       const origin = await this.taskService.getOneTaskById(taskId);
 
@@ -123,13 +186,20 @@ export default class TaskResolver {
     }
   }
 
-  // TODO: 不可删除已指派的任务
   @Mutation(() => TaskStatus, { nullable: false, description: "删除任务" })
-  async DeleteTask(@Arg("taskId") taskId: string): Promise<TaskStatus> {
+  async DeleteTask(@Arg("taskId") taskId: number): Promise<TaskStatus> {
     try {
       const res = await this.taskService.getOneTaskById(taskId);
       if (!res) {
         return new StatusHandler(false, RESPONSE_INDICATOR.NOT_FOUND, []);
+      }
+
+      if (res.assignee) {
+        return new StatusHandler(
+          false,
+          RESPONSE_INDICATOR.MUTATION_NOT_ALLOWED,
+          []
+        );
       }
 
       await this.taskService.deleteTask(taskId);
@@ -140,43 +210,47 @@ export default class TaskResolver {
     }
   }
 
-  // TODO: 将事务操作也移到service中
-  @Transaction()
   @Mutation(() => TaskStatus, {
     nullable: false,
     description: "创建任务同时关联到实体",
   })
   async CreateNewTask(
-    @Arg("taskCreateParam") param: TaskCreateInput,
-    @TransactionRepository(Task)
-    taskTransRepo: Repository<Task>
+    @Arg("taskCreateParam") param: TaskCreateInput
   ): Promise<TaskStatus> {
     try {
       const { substanceId } = param;
       const substance = await this.SubstanceService.getOneSubstanceById(
         substanceId
       );
+
       if (!substance) {
         return new StatusHandler(false, RESPONSE_INDICATOR.NOT_FOUND, []);
       }
       if (substance.relatedTaskId) {
-        // TODO: 专用标识：实体已关联
-        return new StatusHandler(false, RESPONSE_INDICATOR.EXISTED, []);
+        return new StatusHandler(false, RESPONSE_INDICATOR.EXISTED, [
+          substance,
+        ]);
       }
 
       const { taskTitle } = param;
+
       const isTitleUsed = await this.taskService.getOneTaskByConditions(
         {
           taskTitle,
         },
         ["taskSubstance"]
       );
+
       if (isTitleUsed) {
-        return new StatusHandler(false, RESPONSE_INDICATOR.EXISTED, []);
+        return new StatusHandler(false, RESPONSE_INDICATOR.EXISTED, [
+          isTitleUsed,
+        ]);
       }
+
       param.taskSubstance = substance;
 
-      const res = await taskTransRepo.save(param);
+      const res = await this.taskService.createTask(param);
+
       return new StatusHandler(true, RESPONSE_INDICATOR.SUCCESS, [res]);
     } catch (error) {
       return new StatusHandler(false, JSON.stringify(error), []);
@@ -204,22 +278,16 @@ export default class TaskResolver {
     }
   }
 
-  @Transaction()
   @Mutation(() => TaskStatus, { nullable: false, description: "指派任务" })
   async AssignTask(
-    @Arg("taskId") taskId: string,
-    @Arg("uid") uid: string,
-    @TransactionRepository(Task)
-    taskTransRepo: Repository<Task>
+    @Arg("taskId") taskId: number,
+    @Arg("uid") uid: number
   ): Promise<TaskStatus> {
     try {
       const assignee = await this.executorService.getOneExecutorById(uid);
-      if (!assignee) {
-        return new StatusHandler(false, RESPONSE_INDICATOR.NOT_FOUND, []);
-      }
-
       const task = await this.taskService.getOneTaskById(taskId, ["assignee"]);
-      if (!task) {
+
+      if (!task || !assignee) {
         return new StatusHandler(false, RESPONSE_INDICATOR.NOT_FOUND, []);
       }
 
@@ -228,7 +296,7 @@ export default class TaskResolver {
       }
       task.assignee = assignee;
 
-      const assignRes = await taskTransRepo.save(task);
+      const assignRes = await this.taskService.createTask(task);
       return new StatusHandler(true, RESPONSE_INDICATOR.SUCCESS, [assignRes]);
     } catch (error) {
       return new StatusHandler(false, JSON.stringify(error), []);
@@ -240,7 +308,7 @@ export default class TaskResolver {
     description: "变更任务级别",
   })
   async MutateTaskLevel(
-    @Arg("taskId") taskId: string,
+    @Arg("taskId") taskId: number,
     @Arg("level", (type) => DifficultyLevel) level: DifficultyLevel
   ): Promise<TaskStatus> {
     try {
@@ -263,7 +331,7 @@ export default class TaskResolver {
     nullable: false,
     description: "冻结任务 无法恢复",
   })
-  async FreezeTask(@Arg("taskId") taskId: string): Promise<TaskStatus> {
+  async FreezeTask(@Arg("taskId") taskId: number): Promise<TaskStatus> {
     try {
       const res = await this.taskService.getOneTaskById(taskId);
       if (!res) {
